Update
plugins update

moved all non essential assets to a new folder (_ForTesting)  so we know to not use them for the main game

created resources folder for base models, materials, shaders, soundfx and more.

reworked the player movement and rotation:
now uses rb to move, added acceleration and deceleration,
can control the rotation amount and speed for each direction and movement

added aim speed for the player aiming, and made the aiming smooth,
reworked auto center for aiming

reworked the projectile behaviors list, they are not SO anymore
can select for each weapon from a list.